% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-slopeflex.R
\name{plot.slopeflex}
\alias{plot.slopeflex}
\title{plot a slopeflex object}
\usage{
\method{plot}{slopeflex}(sx_obj, ...)
}
\arguments{
\item{sx_obj}{an object generated by `slopeflex_run_model`}
}
\description{
plot a slopeflex object
}
