% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-package.R
\docType{package}
\name{slopeflex-package}
\alias{slopeflex-package}
\alias{slopeflex}
\title{The 'slopeflex' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
