% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-build-model.R
\name{slopeflex_build_model}
\alias{slopeflex_build_model}
\title{function to build a slopeflex model matrix}
\usage{
slopeflex_build_model(ds, sf, h)
}
\arguments{
\item{ds}{data frame or tibble with 2 columns. First column is a date, second some value}

\item{sf}{data frame or tibble with 3 columns.
First column is date, second the expected value of the slope change,
third is the standard deviation of the the slope change.}

\item{h}{positive integer value. this is the forecast horizon.}
}
\description{
function to build a slopeflex model matrix
}
