% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-fit-model.R
\name{slopeflex_fit_model}
\alias{slopeflex_fit_model}
\title{function to build a slopeflex model matrix}
\usage{
slopeflex_fit_model(ds, sf, h, mc.cores = 1, mcmc_list = list(n_iter =
  2500, n_chain = 1, n_thin = 1, n_warmup = 500, control = list(adapt_delta
  = 0.8, max_treedepth = 10)))
}
\arguments{
\item{ds}{data frame or tibble with 2 columns. First column is a date, second some value}

\item{sf}{data frame or tibble with 3 columns.
First column is date, second the expected value of the slope change,
third is the standard deviation of the the slope change.}

\item{h}{positive integer value. this is the forecast horizon.}

\item{mc.cores}{number of cores to use. defaults 1 (recommended; faster)}

\item{mcmc_list}{rstan list of options}
}
\description{
function to build a slopeflex model matrix
}
