% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-fit-model.R
\name{slopeflex_fit_model_lm}
\alias{slopeflex_fit_model_lm}
\title{function to build a slopeflex model matrix using 'lm' as the engine}
\usage{
slopeflex_fit_model_lm(ds, sx, h, lb_ub)
}
\arguments{
\item{ds}{data frame or tibble with 2 columns.
First column is a date, second some value}

\item{sx}{data frame or tibble with 3 columns.
First column is date, second the expected value of the slope change,
third is the standard deviation of the the slope change.}

\item{h}{positive integer value. this is the forecast horizon.
added to the end of the last slopeflex (even if unobserved)}

\item{lb_ub}{numeric(2) vector with lower bound and upper bound.}

\item{mc.cores}{number of cores to use. defaults 1 (recommended; faster)}

\item{mcmc_list}{rstan list of options}
}
\description{
function to build a slopeflex model matrix using 'lm' as the engine
}
