% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-plot-fit.R
\name{slopeflex_plot_fit}
\alias{slopeflex_plot_fit}
\title{plot a slopeflex object}
\usage{
slopeflex_plot_fit(fit)
}
\arguments{
\item{fit}{an rstan object generated by `slopeflex_run_model` with `rstan` engine.}
}
\description{
plot a slopeflex object
}
