% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-build-model.R
\name{slopeflex_build_model}
\alias{slopeflex_build_model}
\title{function to build a slopeflex model matrix}
\usage{
slopeflex_build_model(
  ds,
  sx,
  h,
  prior_intercept = c(0, 100),
  prior_slope = c(0, 10)
)
}
\arguments{
\item{ds}{data frame or tibble with 2 columns. First column is a date, second some value}

\item{sx}{data frame or tibble with 3 columns.
First column is date, second the expected value of the slope change,
third is the standard deviation of the the slope change.}

\item{h}{positive integer value. this is the forecast horizon.}

\item{prior_intercept}{numeric(2) with mean and std dev of intercept prior}

\item{prior_slope}{numeric(2) with mean and std dev of initial slope prior}
}
\description{
function to build a slopeflex model matrix
}
