% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeflex-fit-model.R
\name{slopeflex_fit_model}
\alias{slopeflex_fit_model}
\title{function to build a slopeflex model matrix. can handle unobserved slopeflexes
    if engine = 'stan'}
\usage{
slopeflex_fit_model(
  ds,
  sx,
  h,
  engine = c("stan", "lm")[1],
  lb_ub = c(0.05, 0.95),
  ...
)
}
\arguments{
\item{ds}{data frame or tibble with 2 columns.
First column is a date, second some value}

\item{sx}{data frame or tibble with 3 columns.
First column is date, second the expected value of the slope change,
third is the standard deviation of the the slope change.}

\item{h}{positive integer value. this is the forecast horizon.
added to the end of the last slopeflex (even if unobserved)}

\item{engine}{c('stan', 'lm')}

\item{lb_ub}{numeric(2) vector with lower bound and upper bound.}
}
\description{
function to build a slopeflex model matrix. can handle unobserved slopeflexes
    if engine = 'stan'
}
